Create Student Height Converter CM Inches to Feet Inches
Why (int) is used:
(int)(totalInches / 12): This converts the total number of inches into feet. Since a foot is 12 inches, we divide the total inches by 12. The result of this division might have a fractional part (e.g., 8.5 feet). By casting it to an int, we drop the decimal part and only keep the whole number part (8 feet).

Similarly, in the inches calculation, we are left with a remainder after extracting the full feet, and there's no need to cast it since the remaining inches will be displayed as a floating-point number.